Split into one file per class
